// frontend/src/app/core/api/model/teacher-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RefTeacherSpecialtySimpleResponse } from './ref-teacher-specialty-simple-response';


export interface TeacherResponse { 
    id?: number;
    firstName?: string;
    lastName?: string;
    firstNameAr?: string;
    lastNameAr?: string;
    email?: string;
    phone?: string;
    cin?: string;
    gender?: TeacherResponse.GenderEnum;
    code?: string;
    contractType?: TeacherResponse.ContractTypeEnum;
    hireDate?: string;
    status?: TeacherResponse.StatusEnum;
    degree?: TeacherResponse.DegreeEnum;
    specialties?: Set<RefTeacherSpecialtySimpleResponse>;
}
export namespace TeacherResponse {
    export const GenderEnum = {
        Male: 'MALE',
        Female: 'FEMALE'
    } as const;
    export type GenderEnum = typeof GenderEnum[keyof typeof GenderEnum];
    export const ContractTypeEnum = {
        Cdi: 'CDI',
        Cdd: 'CDD',
        Freelance: 'FREELANCE',
        Guest: 'GUEST'
    } as const;
    export type ContractTypeEnum = typeof ContractTypeEnum[keyof typeof ContractTypeEnum];
    export const StatusEnum = {
        Onboarding: 'ONBOARDING',
        Active: 'ACTIVE',
        OnLeave: 'ON_LEAVE',
        Suspended: 'SUSPENDED',
        Departed: 'DEPARTED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
    export const DegreeEnum = {
        Bachelor: 'BACHELOR',
        FourYearBachelor: 'FOUR_YEAR_BACHELOR',
        Master: 'MASTER',
        EngineeringDegree: 'ENGINEERING_DEGREE',
        Doctorate: 'DOCTORATE'
    } as const;
    export type DegreeEnum = typeof DegreeEnum[keyof typeof DegreeEnum];
}